refactor: updates latest prices layout
